replace setting pages logic into the pagination component
